requirements.txt
fastapi
sqlalchemy
python-jose
passlib
pydantic
bcrypt